improve scripts and experiment with dawn's cmake settings
